requirements.txt
fastapi
uvicorn
streamlit
clickhouse-connect
requests
pandas
pydantic
python-dotenv
plotly
